Scaladoc is now pointing to sources in github
For snapshots, it points to the exact commit, for releases it points to
the tag ("v" + maven version). The link now opens in a different tab,
as opening in the same frame is not compatible with github (the page
doesn't load for some reason).

Left the repo url in test/review untouched because it points to the
root of all LAMP repos. But... is anyone still using that script?!?