Do not include partest and it's dependencies in generated pom.
Define custom Ivy configuration that contains all dependencies of
partest (and partest itself) and make sure that this configuration
is not considered when pom is generated.

Remove the comment about significance of dependency order declaration.
Given the fact that we excluded conflicting, transitive dependencies
and all partest dependencies are isolated from regular dependencies
the situation is much better.