Fix #51 Test compat for -Xxml:coalescing
Rather than maintain a test to compile under multiple
regimes, just build under `-Xxml:coalescing`.

A few other flags are supplied. Deprecation is not yet
possible because of `WordExp` et al. Renamed test classes
`A` and `B` because sometimes it's nice to have a type
parameter called `A` or `B`.