Move Google copyright to file header
The copyright should just be in the file header for the Elem and
MetaData objects and classes, and not in their documentation.

Currently, these copyright notices are showing up in the HTML api docs
(scaladoc) for those elements.  The rest of the library doesn't do
this, so it's probably worth moving it out, rather than pasting
copyright notices everywhere in the api docs for consistency.

According to the original author of scala-xml, Burak Emir:

  The patch was built in 20% time while I was already working for
  Google, using Google's resources. The patch was formally reviewed
  and approved and Google is ok with it being part of Scala and people
  using, distributing, etc... it under the Scala license.

  That is why the copyright notice is there. See "committed to open
  source" in
  https://abc.xyz/investor/other/google-code-of-conduct.html